requirements.txt
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
openai==1.3.0
google-generativeai==0.3.0
aiohttp==3.8.5
python-dotenv==1.0.0
requests==2.31.0